restructure: restructured approvals page to work with new hook
