+++ app/frontend/src/pages/admin/approvals.tsx
@@ -1,97 +1,29 @@
 import React, { useState, useEffect } from 'react';
 import type { RoleRequest,RoleStatus } from '../../types/role-request';
+import { useRoleRequests } from '../../hooks/useRoleRequests';
 import { RoleApprovalModal } from '../../components/admin/approvalModal';
 import { AdminSideBar } from '../../components/admin/adminSidebar';
 import { RoleFilterTabs } from '../../components/admin/approvalFilter';
 import { RoleRequestsTable } from '../../components/admin/approvalTable';
 
 export default function RoleApprovalPage() {
-    const [request, setRequest] = useState<RoleRequest[]>([]);
+    const { requests, loading, approveRequest, rejectRequest, revokeRequest } = useRoleRequests();
     const [selectedRequest, setSelectedRequest] = useState<RoleRequest | null>(null);
     const [filter, setFilter] = useState<'All' | RoleStatus>('All');
-    const [loading, setLoading] = useState(true);
-
-    useEffect(() => {
-        const fetchRequest = async() => {
-            const url = `/api/admin/role-requests`;
-            console.log('Fetching from:', url);
-            try{
-                const resp = await fetch(`/api/admin/role-requests`);
-                if (!resp.ok) {
-                    console.warn("API unavailable");
-                    setRequest([])
-                    return;
-                }
-                const data = await resp.json();
-                setRequest(data.data ?? []);
-            }catch (error){
-                console.error("Failed to load role requests", error);
-                setRequest([])
-            }finally{
-                setLoading(false);
-            }
-        }
-        fetchRequest();
-    }, []);
 
     const handleApprove = async (requestId: string) => {
-        try{
-            const resp = await fetch(`/api/admin/role-requests/${requestId}/approve`, {
-                method: 'PUT'
-            });
-
-            if(resp.ok){
-                const updateRequest = await resp.json();
-
-                setRequest(prev => prev.map(req =>
-                    req.request_id === requestId ? updateRequest : req
-                ));
-
-                setSelectedRequest(null); // close modal
-            }else{
-                console.error("Approval failed");
-            }
-        }catch(error){
-            console.error("Error approving request");
-        }
-    }
+        await approveRequest(requestId);
+        setSelectedRequest(null);
+    };
 
     const handleReject = async(requestId: string) => {
-        try{
-            const resp = await fetch(`/api/admin/role-requests/${requestId}/reject`, {
-                method: 'PUT'
-            });
-
-            if(resp.ok){
-                setRequest(prev => prev.map(req =>
-                    req.request_id === requestId ? { ...req, status: 'rejected' as RoleStatus } : req
-                ));
-                setSelectedRequest(null);
-            }else{
-                console.error("Reject failed:", await resp.text());
-            }
-        }catch(error){
-            console.error("Error rejecting request", error);
-        }
+        await rejectRequest(requestId);
+        setSelectedRequest(null);
     };
 
     const handleRevoke = async(requestId: string) => {
-        try{
-            const resp = await fetch(`/api/admin/role-requests/${requestId}/revoke`, {
-                method: 'PUT',
-            });
-
-            if(resp.ok){
-                setRequest(prev => prev.map(req =>
-                    req.request_id === requestId ? {...req, status: 'revoked' as RoleStatus} : req
-                ));
-                setSelectedRequest(null)
-            }else{
-                console.error("Revoke failed:", await resp.text());
-            }
-        }catch(error){
-            console.error("Error revoking request", error);
-        }
+        await revokeRequest(requestId);
+        setSelectedRequest(null);
     };
 
     if (loading) {
@@ -115,7 +47,7 @@ export default function RoleApprovalPage() {
                 <RoleFilterTabs filter={filter} onChange={setFilter}/>
 
                 {/* table */}
-                <RoleRequestsTable requests={request} filter={filter} onView={setSelectedRequest} />
+                <RoleRequestsTable requests={requests} filter={filter} onView={setSelectedRequest} />
 
                 {/* modal overlay */}
                 {selectedRequest && (
