+++ app/frontend/src/pages/users/live-map.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { NearbyReports, useNearbyFires} from '../../components/nearbyReports';
+import { NearbyReports, useNearbyFires} from '../../components/shared/nearbyReports';
 import { UserSideBar } from '../../components/users/usersSidebar';
 import { PageHeader } from '../../components/layout/pageHeader';
 import { MapPanel } from '../../components/users/mapPanel';
